+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # Change Log
 
+#### 1.3.1
+- Changed the browser to open at lease one instance always. 
+
 #### 1.3.0
 - Added the ability to log to files and display them.
 - Added the `process` route.